Time Conversion
In this program, the user inputs the time in seconds (including decimals), and gets the time converted into various units of time. NOTE: There is a slight error in the program. If you know the solution, I humbly request you to suggest the correct version, ASAP. Hope you appreciate it.